# spec/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AgentAPI Proxy",
    "description": "API for managing agent sessions, schedules, and settings. AgentAPI Proxy provides a unified interface for running AI agents with Kubernetes-based session management.",
    "version": "1.0.0",
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
  "paths": {
...
    "/start": {
      "post": {
        "summary": "Create a new session",
        "description": "Creates a new session and starts an agentapi server instance. Returns a stable public session ID for subsequent requests. When allocation selects a different concrete session ID, the returned ID remains a routable alias for the lifetime of the session.",
        "operationId": "createSession",
        "tags": [
          "Sessions"
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartRequest"
              },
              "examples": {
                "basic": {
                  "summary": "Basic session",
                  "value": {
                    "tags": {
                      "repository": "org/repo",
                      "branch": "main"
                    }
                  }
                },
                "with-environment": {
                  "summary": "Session with environment variables",
                  "value": {
                    "environment": {
                      "CUSTOM_VAR": "value"
                    },
                    "tags": {
                      "repository": "org/repo"
                    }
                  }
                },
                "with-params": {
                  "summary": "Session with initial message",
                  "value": {
                    "tags": {
                      "repository": "org/repo"
                    },
                    "params": {
                      "message": "Review the latest PRs"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "session_id": {
                      "type": "string",
                      "description": "Stable public session identifier (UUID). This ID remains valid even when the concrete allocated session has a different ID."
                    }
                  },
                  "required": [
                    "session_id"
                  ]
                },
                "example": {
                  "session_id": "abc123-def456-ghi789"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Failed to create session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/sessions/{sessionId}": {
      "delete": {
        "summary": "Delete a session or pending session allocation",
        "description": "Terminates and deletes a session, or cancels a pending session allocation. Users can only delete their own sessions and allocations.",
        "operationId": "deleteSession",
        "tags": [
          "Sessions"
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session or pending session allocation deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Session terminated successfully"
                    },
                    "session_id": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "example": "terminated"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Session ID is required"
          },
          "403": {
            "description": "Forbidden - can only delete own sessions"
          },
          "404": {
            "description": "Session not found"
          },
          "409": {
            "description": "Session allocation is no longer pending"
          },
          "500": {
            "description": "Failed to delete session"
          }
        }
      }
    },
...
    "/sessions/status/stream": {
      "get": {
        "summary": "Stream all session status changes (SSE)",
        "description": "Opens a Server-Sent Events stream that pushes a `SessionStatusEvent` whenever any session accessible to the authenticated user changes status. A heartbeat comment (`: heartbeat`) is sent every 30 seconds to keep the connection alive. The stream stays open until the client disconnects.",
        "operationId": "streamSessionsStatus",
        "tags": [
          "Sessions"
        ],
        "responses": {
          "200": {
            "description": "SSE stream of SessionStatusEvent objects. Each event line: `data: {\"session_id\":\"...\",\"status\":\"...\",\"timestamp\":\"...\"}`",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                },
                "example": "data: {\"session_id\":\"abc123\",\"status\":\"active\",\"timestamp\":\"2026-05-02T12:00:00Z\"}\n\n"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "501": {
            "description": "Not implemented for this session manager type"
          }
        }
      }
    },
    "/sessions/status/wait": {
      "get": {
        "summary": "Long-poll next session status change",
        "description": "Blocks until any session accessible to the authenticated user changes status, or until the timeout expires. Returns a `SessionStatusEvent` on change, or `{\"events\": []}` on timeout.",
        "operationId": "waitSessionsStatus",
        "tags": [
          "Sessions"
        ],
        "parameters": [
          {
            "name": "timeout",
            "in": "query",
            "description": "Maximum wait time in seconds (default: 30, max: 60).",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 60,
              "default": 30
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A `SessionStatusEvent` if a status change occurred before the timeout, or `{\"events\": []}` if the timeout elapsed with no change.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/SessionStatusEvent"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "events": {
                          "type": "array",
                          "items": {},
                          "maxItems": 0
                        }
                      },
                      "required": [
                        "events"
                      ]
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "501": {
            "description": "Not implemented for this session manager type"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "StartRequest": {
        "type": "object",
        "properties": {
          "environment": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Environment variables for the session"
          },
          "tags": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Tags for the session (e.g., repository, branch). Keys prefixed with allocator. select an External Session Manager by exact label match; multiple allocator.* tags use AND semantics. allocator.id matches the manager ID."
          },
          "params": {
            "$ref": "#/components/schemas/SessionParams"
          },
          "scope": {
            "$ref": "#/components/schemas/ResourceScope",
            "description": "Ownership scope of the session (default: user)"
          },
          "team_id": {
            "type": "string",
            "description": "Team identifier (e.g., 'org/team-slug') when scope is 'team'"
          },
          "memory_key": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Custom tag map to identify memories for this session. If not specified or empty, the session tags are used as the memory key. If both are empty, memory integration is disabled.",
            "example": {
              "project": "myapp",
              "env": "production"
            }
          }
        }
      },
      "SessionParams": {
        "type": "object",
        "properties": {
          "resume_from": {
            "type": "string",
            "description": "Previous proxy session ID whose persisted ACP conversation state should be restored before starting the agent"
          },
          "message": {
            "type": "string",
            "description": "Initial message to send to the agent after session starts"
          },
          "github_token": {
            "type": "string",
            "description": "GitHub token to use for authentication instead of GitHub App"
          },
          "agent_type": {
            "type": "string",
            "description": "Agent type for the session. Supported values include 'auto', 'claude-acp', 'codex-acp', and 'cursor'. 'auto' selects codex-acp when Codex auth.json is available, otherwise claude-acp. If not specified, uses default agentapi.",
            "example": "claude-acp"
          },
          "slack": {
            "$ref": "#/components/schemas/SlackParams"
          },
          "oneshot": {
            "type": "boolean",
            "description": "When true, the session will automatically delete itself after Claude stops responding. Uses Claude Code Stop hook to execute session deletion.",
            "default": false
          },
          "initial_message_wait_second": {
            "type": "integer",
            "description": "Number of seconds to wait after the agent becomes stable before sending the initial message. Defaults to 2 seconds if not specified.",
            "default": 2,
            "example": 5
          },
          "cycle_message": {
            "type": "string",
            "description": "Message to send to the session after each Claude stop event. When set, a Stop hook is injected into the session's Claude settings that runs `agentapi-proxy client cycle <message>`. The cycle continues until /tmp/check/CYCLE_OK exists or cycle_max_count is reached.",
            "example": "Please continue the task"
          },
          "cycle_max_count": {
            "type": "integer",
            "description": "Maximum number of cycles to run. 0 means unlimited. Requires cycle_message to be set. The current count is tracked in /tmp/check/CYCLE_COUNT inside the session.",
            "default": 10,
            "example": 10
          },
          "sandbox": {
            "$ref": "#/components/schemas/SandboxParams"
          },
          "docker": {
            "$ref": "#/components/schemas/DockerParams"
          },
          "session_ttl": {
            "type": "string",
            "description": "Duration after the last message before this session is automatically deleted. Accepted format: Go duration string (e.g. '48h', '168h', '7d' is not valid — use hours). Empty string means the global cleanup worker TTL is used for Slackbot sessions; non-Slackbot sessions without this field are not auto-deleted.",
            "example": "48h"
          },
          "unsynced_file_paths": {
            "type": "array",
            "description": "Managed file paths that should not be synced back to storage from this session.",
            "items": {
              "type": "string"
            },
            "example": [
              "/home/agentapi/.codex/auth.json"
            ]
          },
          "credential_source": {
            "type": "string",
            "enum": [
              "session_user",
              "triggered_user",
              "team",
              "none"
            ],
            "description": "Managed credential files to inject. session_user uses the trigger configuration owner, triggered_user uses the rendered username tag and falls back to team credentials, team uses the session team, and none disables credential injection. When omitted, user-scoped sessions use session_user and team-scoped sessions use none.",
            "example": "session_user"
          }
        }
      },
      "DockerParams": {
        "type": "object",
        "description": "Docker-in-Docker (DinD) configuration for the session. When enabled, a privileged DinD sidecar is injected into the session Pod and DOCKER_HOST is set so the main container can run docker commands.",
        "properties": {
          "enabled": {
            "type": "boolean",
            "description": "When true, a DinD sidecar (docker:dind) is added to the session Pod and DOCKER_HOST=tcp://127.0.0.1:2375 is set in the main container.",
            "default": false
          },
          "registries": {
            "type": "array",
            "description": "Authenticated container registries. Supports inline credentials (username/password) or a reference to a Kubernetes Secret containing docker config JSON.",
            "items": {
              "$ref": "#/components/schemas/DockerRegistry"
            }
          }
        }
      },
      "DockerRegistry": {
        "type": "object",
        "description": "Authentication configuration for a container registry.",
        "properties": {
          "server": {
            "type": "string",
            "description": "Registry server address (e.g. 'ghcr.io', 'registry.example.com'). Leave empty for Docker Hub.",
            "example": "ghcr.io"
          },
          "username": {
            "type": "string",
            "description": "Registry username for inline credentials. Used together with password."
          },
          "password": {
            "type": "string",
            "description": "Registry password or access token for inline credentials."
          },
          "secret_name": {
            "type": "string",
            "description": "Name of a Kubernetes Secret containing a 'config.json' key with docker config JSON. When set, this secret is mounted into the DinD container for daemon-level auth. Takes priority over cluster-wide dind_registry_secret_name.",
            "example": "my-registry-secret"
          },
          "insecure": {
            "type": "boolean",
            "description": "When true, adds --insecure-registry=<server> to the DinD daemon args, allowing plain HTTP communication with this registry.",
            "example": true
          }
        }
      },
      "SandboxParams": {
        "type": "object",
        "description": "Network sandbox configuration for the session. When enabled, outbound HTTP/HTTPS traffic is routed through a transparent proxy sidecar. allowed_domains (allowlist) takes precedence over denied_domains (denylist). A policy_id may reference a SandboxPolicy resource whose domain lists are merged before session-level overrides.",
        "properties": {
          "enabled": {
            "type": "boolean",
            "description": "When true, the network-filter init container and sidecar are injected into the session Pod. The init container configures iptables to redirect HTTP/HTTPS traffic through the sidecar.",
            "default": false
          },
          "policy_id": {
            "type": "string",
            "description": "ID of a SandboxPolicy resource to use as the base rule set. The policy's allowed_domains and denied_domains are merged with session-level overrides. Session-level domains take precedence.",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "allowed_domains": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Allowlist mode: only listed destinations are permitted; all others are blocked. Supports hostname wildcard prefixes (e.g. '*.example.com'), exact IPv4 addresses, and IPv4 CIDR ranges. IP addresses and CIDR ranges are installed as direct iptables allow rules by the init container. Takes precedence over denied_domains when set.",
            "example": [
              "api.example.com",
              "*.trusted.com",
              "192.0.2.10",
              "198.51.100.0/24"
            ]
          },
          "denied_domains": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Denylist mode: listed domains are blocked, all others are allowed. Used only when allowed_domains is empty. Supports wildcard prefixes (e.g. '*.example.com').",
            "example": [
              "example.com",
              "*.evil.com"
            ]
          },
          "count_mode": {
            "type": "boolean",
            "description": "When true, enables count mode: policy rules are evaluated and blocked domains are recorded, but traffic is not actually blocked. Useful for auditing policies before enforcing them in production. Requires nfa v0.6.0+.",
            "default": false
          }
        }
      },
...
      "SlackParams": {
        "type": "object",
        "properties": {
          "channel": {
            "type": "string",
            "description": "Slack channel ID (e.g., 'C1234567890')",
            "example": "C1234567890"
          },
          "thread_ts": {
            "type": "string",
            "description": "Thread timestamp for threaded messages (e.g., '1234567890.123456')",
            "example": "1234567890.123456"
          }
        },
        "description": "Slack integration parameters for the session"
      },
...
      "SessionStatus": {
        "type": "string",
        "enum": [
          "creating",
          "starting",
          "running",
          "active",
          "unhealthy",
          "stopped",
          "unknown"
        ],
        "description": "Session status. 'running' means the agentapi backend is actively processing a message."
      },
      "SessionStatusEvent": {
        "type": "object",
        "description": "A proxy-level event emitted whenever a session's status changes.",
        "properties": {
          "session_id": {
            "type": "string",
            "description": "The session whose status changed"
          },
          "status": {
            "$ref": "#/components/schemas/SessionStatus"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "When the status change was detected by the proxy"
          }
        },
        "required": [
          "session_id",
          "status",
          "timestamp"
        ]
      },
...
      "ResourceScope": {
        "type": "string",
        "enum": [
          "user",
          "team"
        ],
        "default": "user",
        "description": "Ownership scope. 'user' for personal resources, 'team' for team-shared resources"
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "GitHub OAuth token"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Static API key"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKey": []
    }
  ],
...
}